requirements.txt
altair==5.3.0
annotated-types==0.7.0
app-store-scraper==0.3.5
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.7
cryptography==42.0.8
gitdb==4.0.11
google-ai-generativelanguage==0.6.5
google-api-core==2.19.1
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==2.10
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
proto-plus==1.24.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
requests==2.27.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
six==1.16.0
smmap==5.0.1
soupsieve==2.5
streamlit==1.36.0
streamlit-cookies-manager==0.2.0
tenacity==8.4.2
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.25.11
watchdog==4.0.1
